Backend/servicios-vision-artificial/imagen-matriz-api-ms/requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Image Processing
pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.7.0

# Utilities
python-dotenv>=1.0.0
pytest>=7.3.1
httpx>=0.24.0
